refactor: split codegen submodule into subpackage
