+++ src/tubeulator/codegen/gen/__init__.py
@@ -1,0 +1,4 @@
+"""Give access to the DTO emission function."""
+from .main import emit_deserialisers
+
+__all__ = ["emit_deserialisers"]
